invoice-extraction-ai-project/requirements.txt
streamlit
pytesseract
Pillow
openai
supabase
pandas